Modification of the description of a geographical validity defined with areas/zones print statement from 'Hello' to 'Goodbye'
1/ I harmonized the names:
The previous version used 'area' in some elements, and 'zone' in other elements. 
Now all the ASN.1 elements and types now use the word 'area',
and all the comments use "area/zone" to show that both words refer to the same concept.

2/ I integrated an optional code table in AreaIdType type:
This modification copies what has already been done for StationSetType.
This allows to use a different code table for every part of AreaType (restrictedEntry:area, area:included, restrictedExit:area)
(and if this is not needed, the optional code table element is omitted and the code table of the container object applies)

3/ I added OPTIONAL for the 'area' element of AreaType.
This allows a new use case: when this element is omitted, there are not restrictions on how to go from restrictedEntry to restrictedExit.

And finally, I did NOT add a 'sameAsEntry NULL' choice in restrictedExit, because the fact that the direction is - or is not - mandatory is already covered by the element directionIsMandatory in ConditionsType. Indeed, we have the same situation for OriginDestinationType and LineType:
- when DirectionIsMandatory = TRUE, the traveler has to go from O to D, or has to enter the allowed lines in entryStation and leave them in terminatingStation
- when DirectionIsMandatory = FALSE, the traveler can go from O to D or from D to O, or can enter and leave the allowed lines in entryStation or in terminatingStation